refactor(templates): Remove message_contents own controller and actions
