+++ app/controllers/configurations_controller.rb
@@ -1,7 +1,9 @@
 class ConfigurationsController < ApplicationController
   PERMITTED_PARAMS = [
     { invitation_formats: [] }, :convene_applicant, :rdv_with_referents, :file_configuration_id,
-    :invite_to_applicant_organisations_only, :number_of_days_before_action_required, :motif_category_id
+    :invite_to_applicant_organisations_only, :number_of_days_before_action_required, :motif_category_id,
+    :template_rdv_title_override, :template_rdv_title_by_phone_override, :template_rdv_purpose_override,
+    :template_applicant_designation_override
   ].freeze
 
   include BackToListConcern
