+++ app/views/common/_attribute_input.html.erb
@@ -45,6 +45,10 @@
   <% elsif local_assigns[:as] == :boolean %>
     <p><%= f.input attribute, as: :radio_buttons %></p>
   <% else %>
-    <p><%= f.input attribute, input_html: { value: local_assigns[:value] }  %></p>
+    <% if local_assigns[:value] %>
+      <p><%= f.input attribute, input_html: { value: local_assigns[:value] } %></p>
+    <% else %>
+      <p><%= f.input attribute %></p>
+    <% end %>
   <% end %>
 </div>
