+++ app/views/configurations/edit.html.erb
@@ -1,4 +1,4 @@
 <%= simple_form_for(@configuration, url: organisation_configuration_path(@organisation, @configuration), html: { method: :patch }) do |f| %>
-  <%= render 'configuration_form', f: f, title: "Modifier configuration", configuration: @configuration, return_path: organisation_configurations_path(@organisation), current_file_configuration: @file_configuration, organisation: @organisation, template: @template %>
+  <%= render 'configuration_form', f: f, title: "Modifier configuration", configuration: @configuration, return_path: organisation_configuration_path(@organisation, @configuration), current_file_configuration: @file_configuration, organisation: @organisation, template: @template %>
 <% end %>
