+++ app/views/configurations/messages_contents/_messages_content.html.erb
@@ -8,28 +8,20 @@
     <% end %>
   </div>
 
-  <div class="d-flex justify-content-between mb-4 my-4">
-    <div class="d-flex justify-content-between">
-      <div class="mx-1">
-        <%= link_to previews_invitations_path(configuration_id: configuration.id), data: { turbo_frame: 'remote_modal' } do %>
-          <%= button_tag class: "btn btn-blue" do %>
-            Invitations <i class="fas fa-eye"></i>
-          <% end  %>
-        <% end %>
-      </div>
-      <div class="mx-1">
-        <%= link_to previews_notifications_path(configuration_id: configuration.id), data: { turbo_frame: 'remote_modal' } do %>
-          <%= button_tag class: "btn btn-blue" do %>
-            Notifications <i class="fas fa-eye"></i>
-          <% end  %>
-        <% end %>
-      </div>
+  <div class="d-flex justify-content-center mb-4 my-4">
+    <div class="mx-1">
+      <%= link_to previews_invitations_path(configuration_id: configuration.id), data: { turbo_frame: 'remote_modal' } do %>
+        <%= button_tag class: "btn btn-blue" do %>
+          Invitations <i class="fas fa-eye"></i>
+        <% end  %>
+      <% end %>
     </div>
-
-    <%= link_to edit_configuration_messages_content_path(configuration) do %>
-      <%= button_tag class: "btn btn-blue" do %>
-        Modifier
-      <% end  %>
-    <% end %>
+    <div class="mx-1">
+      <%= link_to previews_notifications_path(configuration_id: configuration.id), data: { turbo_frame: 'remote_modal' } do %>
+        <%= button_tag class: "btn btn-blue" do %>
+          Notifications <i class="fas fa-eye"></i>
+        <% end  %>
+      <% end %>
+    </div>
   </div>
 <% end %>
