--- app/views/configurations/messages_contents/show.html.erb
@@ -1,1 +1,0 @@
-<%= render "messages_content", configuration: @configuration, template: @template %>
